revisión clases

+++ src/main/java/com/example/zonafit/mapper/PaymentMapper.java
@@ -4,13 +4,15 @@ import com.example.zonafit.domain.model.Payment;
 import com.example.zonafit.dto.PaymentResponseDTO;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class PaymentMapper {
     
     public PaymentResponseDTO toResponseDTO(Payment payment) {
         return new PaymentResponseDTO(
                 payment.getId(),
-                (double) payment.getAmount(),
+                payment.getAmount(),
                 payment.getPaymentDate(),
                 payment.getPaymentMethod(),
                 payment.getUser().getId(),
